# MIRI says `reverse` is UB, so replace it with something LLVM can vectorize