--- a/src/test/codegen/slice-reverse.rs
+++ b/src/test/codegen/slice-reverse.rs
@@ -1,5 +1,6 @@
 // compile-flags: -O
 // only-x86_64
+// ignore-debug: the debug assertions in from_raw_parts get in the way
 
 #![crate_type = "lib"]