@scottmcm 2021-11-14
I'm quite confident in the `unsafe` in here, so this is scary
```
rustc exited with signal: 11 (core dumped)
error: could not compile `bitflags`
```

I'll just close my eyes and
@bors retry