@Mark-Simulacrum 2021-11-14
This is https://github.com/rust-lang/rust/issues/90812, almost certainly.