@rust-log-analyzer 2021-11-14
A job failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/runs/4205591560?check_suite_focus=true) [(plain)](https://github.com/rust-lang-ci/rust/commit/9e811bc4c0e96a0a804668a9f743d510ef7f2b2a/checks/4205591560/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>